pyproject.toml
[tool.poetry]
name = "lyngdorf"
version = "0.2.10"
description = "Library to control a Lyngdorf A/V processor"
authors = ["Alex Fishlock <alex.fishlock@racingjag.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",      # Chose either "3 - Alpha", "4 - Beta" or "5 - Production/Stable" as the current state of your package
    "Intended Audience :: Developers",      # Define that your audience are developers
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/fishloa/lyngdorf"
Issues = "https://github.com/fishloa/lyngdorf/issues"

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
